Refactor ILIAS link classes
The classes to get an url to ILIAS did not work properly. Therefore, the classes are completly refactored
and changed. Now, a ILIAS link builder class builds an ILIAS link. The builder only be used, when a url should
be opened. This enables the possibility to use the original url as a fallback.